Split `Node` and `EmptyView` from `RootView`
